requirements.txt
amqp==2.5.0
anyjson==0.3.3
asgiref==3.2.3
billiard==3.6.3.0
celery==4.3.0
certifi==2020.4.5.1
chardet==3.0.4
Django==2.0.13
djangorestframework==3.9.4
idna==2.5
importlib-metadata==1.1.3
kombu==4.5.0
pytz==2019.3
redis==3.3.11
requests==2.21.0
sqlparse==0.3.1
urllib3==1.21.1
vine==1.3.0
zipp==1.2.0